% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Popup Label for Earthquake Data}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{- a data frame}
}
\value{
A character vector with HTML formatted labels for each row in the data frame.
}
\description{
eq_create_label generates an HTML formatted text that cotain information about an earthquake event:
 * City
 * Magnitude
 * Number of deaths
}
\note{
Data frame must contain columns for city (CITY), magnitude (Mag), and deaths (Deaths).
}
\examples{


data <- data.frame(
  CITY = c("San Francisco", NA, "Tokyo"),
  Mag = c(6.7, NA, 8.3),
  Deaths = c(3, NA, 125)
)

labels <- eq_create_label(data)

}
