% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Strip Location Names in NOAA Significant Earthquakes dataset}
\usage{
eq_location_clean(d)
}
\arguments{
\item{d}{- a NOAA dataset as data frame.}
}
\value{
The modified data frame (`d`) with a modified dataset.
}
\description{
eq_location_clean cleans and formats location names in a data frame. The names in orignal dataset cotain Country and City in a singe cell.
}
\details{
This function cleans location names in the `"Location Name"` column, assuming the format might include a city/region followed by a colon (":") and a country name. It performs the following steps:
}
\note{
Data frame must contain a column named `"Location Name"` with potentially country-suffixed locations.
This function assumes the `"Location Name"` column might have locations with or without colons separating city/region and country. It uses the `stringr` package for string manipulation functions.
}
\examples{
\dontrun{ clean_data <- eq_location_clean(data) }

}
\seealso{
`strsplit`, `str_trim`, `str_to_title`, `tolower` (from stringr package)
}
