% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{eq_map}
\alias{eq_map}
\title{Create Leaflet Map for Earthquake Data}
\usage{
eq_map(dane, annot_col)
}
\arguments{
\item{dane}{- A data frame}

\item{annot_col}{-  The name of the column in `dane` containing text for marker popups.}
}
\value{
A Leaflet map object.
}
\description{
eq_map generates a Leaflet map visualizing earthquake data.
}
\note{
containing earthquake data, including columns for longitude (`Longitude`), latitude (`Latitude`), magnitude (`Mag`), and an annotation column specified by `annot_col`.
}
\examples{

\dontrun{ d \%>\% dplyr::filter(COUNTRY == "Mexico" & lubridate::year(Date) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% eq_map("popup_text") }

}
