% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Implementation of the GeomTimeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimeline
}
\description{
A main implementation of the genom
}
\keyword{datasets}
