% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Implementation of the GeomTimelineLabel}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomTimelineLabel
}
\description{
A main implementation of the genom
}
\keyword{datasets}
